[x]: Update due to tests #29
